chore: stage 1 refactor
